fix warning

+++ WDL/IPlug/IControl.cpp
@@ -117,7 +117,7 @@ void IControl::PromptUserInput()
 		else // text entry
 		{
 			int cX = (int) mRECT.MW(), cY = (int) mRECT.MH();
-			IRECT txtRECT = IRECT(cX - (float) PARAM_EDIT_W/2., cY - (PARAM_EDIT_H / 2.), cX + (float) PARAM_EDIT_W/2., cY + (PARAM_EDIT_H / 2.)  ); 
+			IRECT txtRECT = IRECT(cX - PARAM_EDIT_W/2, cY - (PARAM_EDIT_H / 2), cX + PARAM_EDIT_W/2, cY + (PARAM_EDIT_H / 2)  ); 
 			mPlug->GetGUI()->PromptUserInput(this, mPlug->GetParam(mParamIdx), &txtRECT );
 		}
